feat(core): added import job representation object
